2018-04-04 | V1.1 Changelog
BUG FIX: You can now enter the same number as width and height without crashing the page. This was done by adding some code to the loop function that renders the table. It will now loop no more than 50 times.

NEW FEATURE: There is now a Portrait/Landscape toggle that allows users to state the orientation of their document, automatically updating the table to match. The toggle is pure CSS.